+++ pom.xml
@@ -2,10 +2,10 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>alexa-skills-kit-samples</groupId>
-  <artifactId>colorpicker</artifactId>
+  <artifactId>Soup-IT</artifactId>
   <packaging>jar</packaging>
   <version>1.0</version>
-  <name>colorpicker</name>
+  <name>Soup-IT</name>
   <url>http://developer.amazon.com/ask</url>
   <licenses>
     <license>
