+++ src/main/java/soupit/ColorPickerStreamHandler.java
@@ -11,18 +11,18 @@
      the specific language governing permissions and limitations under the License.
 */
 
-package main.java.colorpicker;
+package main.java.soupit;
 
 import com.amazon.ask.Skill;
 import com.amazon.ask.SkillStreamHandler;
 import com.amazon.ask.Skills;
-import main.java.colorpicker.handlers.FallbackIntentHandler;
-import main.java.colorpicker.handlers.HelpIntentHandler;
-import main.java.colorpicker.handlers.LaunchRequestHandler;
-import main.java.colorpicker.handlers.SessionEndedRequestHandler;
-import main.java.colorpicker.handlers.WhatsMyColorIntentHandler;
-import main.java.colorpicker.handlers.CancelandStopIntentHandler;
-import main.java.colorpicker.handlers.MyColorIsIntentHandler;
+import main.java.soupit.handlers.FallbackIntentHandler;
+import main.java.soupit.handlers.HelpIntentHandler;
+import main.java.soupit.handlers.LaunchRequestHandler;
+import main.java.soupit.handlers.SessionEndedRequestHandler;
+import main.java.soupit.handlers.WhatsMyColorIntentHandler;
+import main.java.soupit.handlers.CancelandStopIntentHandler;
+import main.java.soupit.handlers.MyColorIsIntentHandler;
 
 public class ColorPickerStreamHandler extends SkillStreamHandler {
